requirement.txt
Django==2.2
djangorestframework==3.11.0
markdown==3.2.1
django-filter==2.2.0
django-rest-swagger==2.2.0
django-rest-auth==0.9.5
django-allauth==0.41.0
django-cors-headers==3.1.0
django-extensions==2.2.9
rest-condition==1.0.3